+++ lib/mineshaft/constants.rb
@@ -9,5 +9,7 @@
 module Mineshaft
   module Constants
     RUBY_VERSION_STABLE = '2.5.1'
+    GLOBAL_DIR = "#{Dir.home}/.mineshaft"
+    GLOBAL_BIN = "#{GLOBAL_DIR}/bin"
   end
 end